fix(error 500): adding a friendly error and catching cases when trying to access API without DB
