+++ src/repositories/user-repository.ts
@@ -1,6 +1,8 @@
 import { GetCommand } from "@aws-sdk/lib-dynamodb";
+import { ResourceNotFoundException } from "@aws-sdk/client-dynamodb";
 import { docClient } from "../config/dynamodb";
 import { StoredUser } from "../models/user";
+import { TableNotFoundError } from "../models/errors";
 
 // Pega o nome da tabela das variáveis de ambiente
 const TABLE_NAME = process.env.TABLE_NAME!;
@@ -27,6 +29,10 @@ export async function getUserByEmail(
 
     return result.Item as StoredUser | undefined;
   } catch (error) {
+    if (error instanceof ResourceNotFoundException) {
+      console.error(`Tabela ${TABLE_NAME} não encontrada.`);
+      throw new TableNotFoundError(TABLE_NAME);
+    }
     console.error("Erro ao buscar usuário no DynamoDB:", error);
     throw new Error("Erro no repositório de dados");
   }
